+++ TCC/fraction.h
@@ -6,7 +6,7 @@ namespace fractionRoute {
 	std::vector<Route> splitReinsertion(const std::vector<Route>& solution, const Client& client, int forbiddenRouteIndex, bool& success, const AdjacencyCosts& adjacencyCosts);
 
 	std::vector<Route> knaapSackGreedy(std::vector<Route>& solution, const Client& client, const std::vector<int>& leastInsertPos,
-		std::vector<int>& routesIndexWithResidual /*L*/, std::vector<float>& residualsList /*A*/, std::vector<float>& leastInsertionCosts /*U*/, bool& success);
+		std::vector<int>& routesIndexWithResidual /*L*/, std::vector<double>& residualsList /*A*/, std::vector<double>& leastInsertionCosts /*U*/, bool& success);
 
 	std::vector<Route> emptyRoutes(const std::vector<Route>& solution, int maxVels, const AdjacencyCosts& adjacencyCosts);
